feat: add custom url matchers for product and category route
