+++ src/app/pages/app-routing.module.ts
@@ -6,6 +6,8 @@ import { FeatureToggleGuard } from 'ish-core/feature-toggle.module';
 import { AuthGuard } from 'ish-core/guards/auth.guard';
 import { LoginGuard } from 'ish-core/guards/login.guard';
 import { LogoutGuard } from 'ish-core/guards/logout.guard';
+import { categoryRouteMatcher } from 'ish-core/route-formats/category.route';
+import { productRouteMatcher } from 'ish-core/route-formats/product.route';
 
 const routes: Routes = [
   { path: '', redirectTo: '/home', pathMatch: 'full' },
@@ -32,12 +34,12 @@ const routes: Routes = [
     },
   },
   {
-    path: 'product',
+    matcher: productRouteMatcher,
     loadChildren: () => import('./product/product-page.module').then(m => m.ProductPageModule),
     canActivate: [MetaGuard],
   },
   {
-    path: 'category',
+    matcher: categoryRouteMatcher,
     loadChildren: () => import('./category/category-page.module').then(m => m.CategoryPageModule),
     canActivate: [MetaGuard],
   },
